Launch details fragment featured

+++ app/src/main/java/ru/alexmaryin/spacextimes_rx/data/model/extra/Failure.kt
@@ -1,7 +1,7 @@
 package ru.alexmaryin.spacextimes_rx.data.model.extra
 
 data class Failure(
-    val time: Float,
-    val altitude: Float,
+    val time: Int,
+    val altitude: Int,
     val reason: String,
 )
